Getting BibTeX citation from OEIS sequence

Getting BibTeX citation from OEIS sequence

Usage

OEIS_bibtex(x)

Arguments

x

An object of class OEIS_sequence

Value

An object of class bibentry with the citation, the function populates a Misc entry type, for use when nothing else fits, with the following fields:

key: a character string giving the citation key for the OEIS sequence.

author: The name(s) of the author(s) (in the case of more than one author, separated by and).

title: The On-Line Encyclopedia of Integer Sequences.

url: Link to sequence in OEIS.

year: The year of publication (or, if unpublished, the year of creation).

month: The month of publication (or, if unpublished, the month of creation).

note: Sequence description.

Note

Fields tags are separated by commas. The last tag finishes with a comma , although is not necessarily. The tag's name is not case-sensitive but caps are used.

For more help on bibentry class, see: utils::bibentry()

References

BibTeX Format Description

BibTeX AUTOMATIC OEIS CITATIONS

Mathematica OEIS Package

Wikipedia: BibTex

Referencing a Particular Sequence

See also

Examples

if (FALSE) { # \dontrun{
x <- OEIS_sequence("A170401")
bref <- OEIS_bibtex(x)
# BibTeX and citation.
print(bref, style = "Bibtex")
print(bref, style = "citation")
} # }